pdfwrite - fix 32 bit build and XRef streams
Bug #707610 "gs10030w32.exe 32bit version "-sDEVICE=pdfwrite" creates damaged pdf file"

The PRId64 macro doesn't work well when the data type is a 'long' on
Windows, and the word size is 32 bits.

A better way to fix this is to promote the structure members to be
64-bit using the int64_t data type, but that's too invasive a change
while we are in the middle of a release.

For now fix the actual problem by casting the data types to the higher
type.

+++ devices/vector/gdevpdf.c
@@ -3318,9 +3318,9 @@ pdf_close(gx_device * dev)
                 offs_bytes++;
             }
 
-            gs_snprintf(str, sizeof(str), "%"PRId64" 0 obj\n<<\n/Type /XRef\n", pdev->next_id);
+            gs_snprintf(str, sizeof(str), "%"PRId64" 0 obj\n<<\n/Type /XRef\n", (int64_t)pdev->next_id);
             stream_puts(s, str);
-            gs_snprintf(str, sizeof(str), "/Size %"PRId64"\n", pdev->next_id + 1);
+            gs_snprintf(str, sizeof(str), "/Size %"PRId64"\n", (int64_t)(pdev->next_id + 1));
             stream_puts(s, str);
 
             pprintld2(s, "/Root %ld 0 R /Info %ld 0 R\n", Catalog_id, Info_id);
